requirements.txt
asgiref==3.8.1
babel==2.17.0
Django==4.2.20
django-extra-views==0.14.0
django-haystack==3.3.0
django-oscar==3.2.5
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory-boy==3.2.1
Faker==37.1.0
geographiclib==2.0
geopy==2.4.1
packaging==24.2
phonenumbers==9.0.2
pillow==11.1.0
psycopg2-binary==2.9.10
purl==1.6
PyJWT==2.9.0
python-dotenv==1.1.0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
